fix(powerbi): fix field mapping types for v_d_compartiment view

+++ Poliedro.Eds.Domain/DashboardPowerBI/CompartimentDashboardView/Entities/CompartimentDashboardViewEntity.cs
@@ -7,8 +7,8 @@ public class CompartimentDashboardViewEntity
     [Key]
     public int IdCompartment { get; set; }
     public int Number { get; set; }
-    public int Nominal { get; set; }
-    public int Operative { get; set; }
+    public double Nominal { get; set; }
+    public double Operative { get; set; }
     public double Height { get; set; }
     public int IdProduct { get; set; }
     public string ProductName { get; set; } = default!;
